+++ nbproject/project.properties
@@ -36,10 +36,22 @@ dist.jlink.output=${dist.jlink.dir}/Penyewaan-Lapangan
 endorsed.classpath=
 excludes=
 file.reference.flatlaf-3.6.jar=lib\\flatlaf-3.6.jar
+file.reference.mssql-jdbc-12.10.0.jre11-javadoc.jar=lib\\mssql-jdbc-12.10.0.jre11-javadoc.jar
+file.reference.mssql-jdbc-12.10.0.jre11-sources.jar=lib\\mssql-jdbc-12.10.0.jre11-sources.jar
+file.reference.mssql-jdbc-12.10.0.jre11.jar=lib\\mssql-jdbc-12.10.0.jre11.jar
+file.reference.mssql-jdbc-12.10.0.jre8-javadoc.jar=lib\\mssql-jdbc-12.10.0.jre8-javadoc.jar
+file.reference.mssql-jdbc-12.10.0.jre8-sources.jar=lib\\mssql-jdbc-12.10.0.jre8-sources.jar
+file.reference.mssql-jdbc-12.10.0.jre8.jar=lib\\mssql-jdbc-12.10.0.jre8.jar
 includes=**
 jar.compress=false
 javac.classpath=\
-    ${file.reference.flatlaf-3.6.jar}
+    ${file.reference.flatlaf-3.6.jar}:\
+    ${file.reference.mssql-jdbc-12.10.0.jre11-javadoc.jar}:\
+    ${file.reference.mssql-jdbc-12.10.0.jre11-sources.jar}:\
+    ${file.reference.mssql-jdbc-12.10.0.jre11.jar}:\
+    ${file.reference.mssql-jdbc-12.10.0.jre8-javadoc.jar}:\
+    ${file.reference.mssql-jdbc-12.10.0.jre8-sources.jar}:\
+    ${file.reference.mssql-jdbc-12.10.0.jre8.jar}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
@@ -75,7 +87,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=Penyewaan-Lapangan
-main.class=Dashboard
+main.class=GUI.Dashboard
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
