+++ src/ConnectionClass/DatabaseConnection.java
@@ -1,3 +1,5 @@
+package ConnectionClass;
+
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
